backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.54.3
qdrant-client==1.12.0
beautifulsoup4==4.12.3
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.5
langchain-text-splitters==0.3.2
psycopg2-binary==2.9.10
sqlmodel==0.0.22

